+++ src/Umbraco.Web.Common/DependencyInjection/UmbracoBuilderExtensions.cs
@@ -1,13 +1,10 @@
 using System.Data.Common;
 using System.Net.Http.Headers;
 using System.Reflection;
-using System.Runtime.CompilerServices;
-using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.DataProtection.Infrastructure;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc.ApplicationModels;
-using Microsoft.AspNetCore.Server.Kestrel.Core;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
@@ -36,6 +33,7 @@ using Umbraco.Cms.Core.Templates;
 using Umbraco.Cms.Core.Web;
 using Umbraco.Cms.Infrastructure.BackgroundJobs;
 using Umbraco.Cms.Infrastructure.BackgroundJobs.Jobs;
+using Umbraco.Cms.Infrastructure.BackgroundJobs.Jobs.DistributedJobs;
 using Umbraco.Cms.Infrastructure.BackgroundJobs.Jobs.ServerRegistration;
 using Umbraco.Cms.Infrastructure.DependencyInjection;
 using Umbraco.Cms.Infrastructure.HostedServices;
@@ -168,37 +166,6 @@ public static partial class UmbracoBuilderExtensions
         builder.Services.AddUnique<IApplicationShutdownRegistry, AspNetCoreApplicationShutdownRegistry>();
         builder.Services.AddTransient<IIpAddressUtilities, IpAddressUtilities>();
         builder.Services.AddUnique<IPreviewTokenGenerator, UserBasedPreviewTokenGenerator>();
-
-        return builder;
-    }
-
-    /// <summary>
-    ///     Add Umbraco recurring background jobs
-    /// </summary>
-    public static IUmbracoBuilder AddRecurringBackgroundJobs(this IUmbracoBuilder builder)
-    {
-        // Add background jobs
-        builder.Services.AddRecurringBackgroundJob<TempFileCleanupJob>();
-        builder.Services.AddRecurringBackgroundJob<InstructionProcessJob>();
-        builder.Services.AddRecurringBackgroundJob<TouchServerJob>();
-        builder.Services.AddRecurringBackgroundJob<ReportSiteJob>();
-
-        builder.Services.AddSingleton<IDistributedBackgroundJob, WebhookFiring>();
-        builder.Services.AddSingleton<IDistributedBackgroundJob, ContentVersionCleanupJob>();
-        builder.Services.AddSingleton<IDistributedBackgroundJob, HealthCheckNotifierJob>();
-        builder.Services.AddSingleton<IDistributedBackgroundJob, LogScrubberJob>();
-        builder.Services.AddSingleton<IDistributedBackgroundJob, ScheduledPublishingJob>();
-        builder.Services.AddSingleton<IDistributedBackgroundJob, TemporaryFileCleanupJob>();
-        builder.Services.AddSingleton<IDistributedBackgroundJob, WebhookLoggingCleanup>();
-        builder.Services.AddSingleton<IDistributedBackgroundJob, CacheInstructionsPruningJob>();
-        builder.Services.AddSingleton<IDistributedBackgroundJob, LongRunningOperationsCleanupJob>();
-        builder.Services.AddHostedService<DistributedBackgroundJobHostedService>();
-
-        builder.Services.AddSingleton(RecurringBackgroundJobHostedService.CreateHostedServiceFactory);
-        builder.Services.AddHostedService<RecurringBackgroundJobHostedServiceRunner>();
-        builder.Services.AddHostedService<QueuedHostedService>();
-        builder.AddNotificationAsyncHandler<PostRuntimePremigrationsUpgradeNotification, NavigationInitializationNotificationHandler>();
-        builder.AddNotificationAsyncHandler<PostRuntimePremigrationsUpgradeNotification, PublishStatusInitializationNotificationHandler>();
 
         return builder;
     }
